Fix options.require dependency issue #73
Fixes #73.

Changes dashing dependency loading to continue to attempt to load dependencies
past the first dependency if the first dependency was already loaded by some
other widget.

This is done by changing the resources to be objects with a `src` key, to which
we add a `loaded` key (set to `true`) once it's loaded. If the resource
requested is already loaded and there are more options, `loadResource` continues
with the next resource.

This commit also adds a test for the issue, but creating a new widget type in
the `multiple_dashboards.html` javascript. It loads only `d3`, while the Graph
widget loads both `d3` and `rickshaw`. The integration test now checks that both
`d3` and `Rickshaw` are defined.